Move broadcasterMetrics to a separate file
